refactor: drop CeloConsensusTx newtype, use CeloTxEnvelope as SignedTx
Reverts the ~600-line wrapper introduced in 8c1a706, enabled by the
upstream fix in ethereum-optimism/optimism#20382 (shipped in
op-reth/v2.3.1, which main is now pinned to). That PR moves the
effective_tip_per_gas() call to before into_consensus() in
OpPayloadBuilderCtx::execute_best_transactions, so the pool-layer view
(which already exposes native-equivalent fees on CeloPoolTx) is used
to compute the miner tip — no consensus-layer cache needed.

With the local newtype gone, two impls that relied on CeloConsensusTx
being a local type move to the crates where the remaining types are
local (orphan rule):

- FromConsensusTx<CeloTxEnvelope> for CeloTransaction moves to
  celo-alloy-rpc-types behind a new `reth` feature, together with
  CeloTransactionInfo and cip64_effective_gas_price (mirroring
  op-alloy-rpc-types' reth_compat.rs).
- From<Sealed<TxPostExec>> (required by the PayloadBuilderBuilder
  bound since kona-node v1.5.0) moves onto CeloTxEnvelope in
  celo-alloy-consensus.

Closes #152.
