+++ crates/celo-reth/src/lib.rs
@@ -41,7 +41,6 @@ use reth_primitives_traits::{NodePrimitives, SealedBlock, SealedHeader, SignedTr
 pub mod primitives;
 pub mod receipt;
 pub mod receipts;
-pub mod signed_tx;
 
 #[cfg(feature = "std")]
 pub mod node;
