+++ crates/celo-reth/src/primitives.rs
@@ -1,15 +1,10 @@
 //! Celo-specific primitive types for reth.
 
-use crate::{receipt::CeloReceipt, signed_tx::CeloConsensusTx};
+use crate::receipt::CeloReceipt;
+use celo_alloy_consensus::CeloTxEnvelope;
 
 /// Celo-specific signed transaction type.
-///
-/// This is a thin wrapper around
-/// [`CeloTxEnvelope`](celo_alloy_consensus::CeloTxEnvelope) that carries
-/// ephemeral, native-equivalent fee values for CIP-64 transactions so that
-/// the payload builder's miner-fee accounting produces correct results. See
-/// [`CeloConsensusTx`] for details.
-pub type CeloTransactionSigned = CeloConsensusTx;
+pub type CeloTransactionSigned = CeloTxEnvelope;
 
 /// Celo-specific block type.
 pub type CeloBlock = alloy_consensus::Block<CeloTransactionSigned>;
