+++ crates/celo-reth/src/test_utils.rs
@@ -57,8 +57,7 @@ pub(crate) fn make_test_tx(
         },
     );
 
-    let signed: crate::primitives::CeloTransactionSigned = tx.into();
-    let recovered = Recovered::new_unchecked(signed, sender);
+    let recovered = Recovered::new_unchecked(tx, sender);
     let pooled = CeloPooledTransaction::try_from(recovered.into_inner()).unwrap();
     let inner = TestInnerPoolTx::from_pooled(Recovered::new_unchecked(pooled, sender));
     CeloPoolTx::new(inner)
